requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.1.0
plotly>=5.17.0
numpy>=1.24.0
scipy>=1.11.0

# File processing
openpyxl>=3.1.0
chardet>=5.2.0

# AI/ML
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Utilities
requests>=2.31.0
loguru>=0.7.0
python-decouple>=3.8
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development/Testing
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0
ruff>=0.1.0